requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Browser Automation
playwright==1.40.0

# AI/LLM Integration
google-generativeai==0.3.2
langchain-google-genai==0.0.5
langchain==0.1.0
openai==1.6.1  # For DeepSeek API (OpenAI-compatible endpoint)

# Data Validation & Processing
pydantic==2.5.3
pandas==2.1.4

# Computer Vision & Image Processing
Pillow==10.1.0

# Accessibility Testing
axe-playwright-python==0.1.3

# NLP & Sentiment Analysis
# Using Gemini/DeepSeek APIs instead of local models
# Removed: transformers, torch, sentencepiece (save ~3GB of dependencies)

# Report Generation
Jinja2==3.1.3
WeasyPrint==60.2

# Utilities
colorama==0.4.6
tqdm==4.66.1
python-json-logger==2.0.7
